rename: 수정벌 추천 패키지명 수정
