+++ src/main/java/kr/co/webee/application/bee/recommendation/service/BeeRecommendationService.java
@@ -1,19 +1,19 @@
-package kr.co.webee.application.bee.recommend.service;
+package kr.co.webee.application.bee.recommendation.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import kr.co.webee.application.ai.RagSearchOptions;
-import kr.co.webee.domain.bee.recommend.entity.BeeRecommendation;
-import kr.co.webee.domain.bee.recommend.repository.BeeRecommendationRepository;
+import kr.co.webee.domain.bee.recommendation.entity.BeeRecommendation;
+import kr.co.webee.domain.bee.recommendation.repository.BeeRecommendationRepository;
 import kr.co.webee.domain.user.entity.User;
 import kr.co.webee.domain.user.repository.UserRepository;
 import kr.co.webee.infrastructure.ai.AiPromptExecutor;
 import kr.co.webee.infrastructure.ai.PromptTemplateRegistry;
-import kr.co.webee.presentation.bee.recommend.dto.request.BeeRecommendationRequest;
-import kr.co.webee.presentation.bee.recommend.dto.request.UserCropInfoRequest;
-import kr.co.webee.presentation.bee.recommend.dto.response.BeeRecommendationAiResponse;
-import kr.co.webee.presentation.bee.recommend.dto.response.BeeRecommendationCreateResponse;
-import kr.co.webee.presentation.bee.recommend.dto.response.BeeRecommendationDetailResponse;
-import kr.co.webee.presentation.bee.recommend.dto.response.BeeRecommendationListResponse;
+import kr.co.webee.presentation.bee.recommendation.dto.request.BeeRecommendationRequest;
+import kr.co.webee.presentation.bee.recommendation.dto.request.UserCropInfoRequest;
+import kr.co.webee.presentation.bee.recommendation.dto.response.BeeRecommendationAiResponse;
+import kr.co.webee.presentation.bee.recommendation.dto.response.BeeRecommendationCreateResponse;
+import kr.co.webee.presentation.bee.recommendation.dto.response.BeeRecommendationDetailResponse;
+import kr.co.webee.presentation.bee.recommendation.dto.response.BeeRecommendationListResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
