+++ src/main/java/kr/co/webee/domain/bee/recommendation/repository/BeeRecommendationRepository.java
@@ -1,9 +1,8 @@
-package kr.co.webee.domain.bee.recommend.repository;
+package kr.co.webee.domain.bee.recommendation.repository;
 
 import io.lettuce.core.dynamic.annotation.Param;
-import kr.co.webee.domain.bee.recommend.entity.BeeRecommendation;
+import kr.co.webee.domain.bee.recommendation.entity.BeeRecommendation;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
