+++ src/main/java/kr/co/webee/presentation/bee/recommendation/dto/response/BeeRecommendationDetailResponse.java
@@ -1,7 +1,7 @@
-package kr.co.webee.presentation.bee.recommend.dto.response;
+package kr.co.webee.presentation.bee.recommendation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.co.webee.domain.bee.recommend.entity.BeeRecommendation;
+import kr.co.webee.domain.bee.recommendation.entity.BeeRecommendation;
 import kr.co.webee.domain.bee.type.BeeType;
 import lombok.Builder;
