+++ src/main/java/kr/co/webee/presentation/bee/recommendation/dto/response/BeeRecommendationListResponse.java
@@ -1,13 +1,12 @@
-package kr.co.webee.presentation.bee.recommend.dto.response;
+package kr.co.webee.presentation.bee.recommendation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.co.webee.domain.bee.recommend.entity.BeeRecommendation;
+import kr.co.webee.domain.bee.recommendation.entity.BeeRecommendation;
 import kr.co.webee.domain.bee.type.BeeType;
 import kr.co.webee.domain.profile.crop.type.CultivationType;
 import lombok.Builder;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Builder
 @Schema(description = "수정벌 추천 내역 목록 response")
